GameRunner - extra Proxy stuff to a 'HttpProxy.java'
