+++ src/games/strategy/engine/framework/GameRunner.java
@@ -10,12 +10,7 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.net.InetSocketAddress;
-import java.net.Proxy;
-import java.net.ProxySelector;
-import java.net.URI;
 import java.util.Calendar;
-import java.util.List;
 import java.util.Properties;
 import java.util.Map.Entry;
 import java.util.logging.Level;
@@ -27,8 +22,6 @@ import java.util.prefs.Preferences;
 import javax.swing.JOptionPane;
 import javax.swing.SwingUtilities;
 
-import org.apache.commons.httpclient.HostConfiguration;
-
 import games.strategy.debug.ClientLogger;
 import games.strategy.debug.ErrorConsole;
 import games.strategy.engine.ClientContext;
@@ -36,6 +29,7 @@ import games.strategy.engine.ClientFileSystemHelper;
 import games.strategy.engine.framework.lookandfeel.LookAndFeel;
 import games.strategy.engine.framework.map.download.MapDownloadController;
 import games.strategy.engine.framework.startup.ui.MainFrame;
+import games.strategy.engine.framework.system.HttpProxy;
 import games.strategy.engine.lobby.server.LobbyServer;
 import games.strategy.triplea.settings.SystemPreferenceKey;
 import games.strategy.triplea.settings.SystemPreferences;
@@ -63,9 +57,6 @@ public class GameRunner {
   public static final int PORT = 3300;
   public static final String DELAYED_PARSING = "DelayedParsing";
   public static final String CASUALTY_SELECTION_SLOW = "CasualtySelectionSlow";
-  public static final String PROXY_CHOICE = "proxy.choice";
-  public static final String HTTP_PROXYHOST = "http.proxyHost";
-  public static final String HTTP_PROXYPORT = "http.proxyPort";
   // do not include this in the getProperties list. they are only for loading an old savegame.
   public static final String OLD_EXTENSION = ".old";
   // argument options below:
@@ -84,8 +75,6 @@ public class GameRunner {
   public static final String LOBBY_GAME_SUPPORT_PASSWORD = "triplea.lobby.game.supportPassword";
   public static final String LOBBY_GAME_RECONNECTION = "triplea.lobby.game.reconnection";
   public static final String TRIPLEA_ENGINE_VERSION_BIN = "triplea.engine.version.bin";
-  public static final String PROXY_HOST = "proxy.host";
-  public static final String PROXY_PORT = "proxy.port";
   public static final String TRIPLEA_DO_NOT_CHECK_FOR_UPDATES = "triplea.doNotCheckForUpdates";
   // has the memory been manually set or not?
   public static final String TRIPLEA_MEMORY_SET = "triplea.memory.set";
@@ -119,9 +108,6 @@ public class GameRunner {
   public static final String MAP_FOLDER = "mapFolder";
 
 
-  public static enum ProxyChoice {
-    NONE, USE_SYSTEM_SETTINGS, USE_USER_PREFERENCES
-  }
 
   public static void usage(GameMode gameMode) {
     if(gameMode == GameMode.HEADLESS_BOT) {
@@ -160,8 +146,8 @@ public class GameRunner {
           + "   " + LOBBY_HOST + "=<LOBBY_HOST>\n"
           + "   " + LOBBY_GAME_COMMENTS + "=<LOBBY_GAME_COMMENTS>\n"
           + "   " + LOBBY_GAME_HOSTED_BY + "=<LOBBY_GAME_HOSTED_BY>\n"
-          + "   " + PROXY_HOST + "=<Proxy_Host>\n"
-          + "   " + PROXY_PORT + "=<Proxy_Port>\n"
+          + "   " + HttpProxy.PROXY_HOST + "=<Proxy_Host>\n"
+          + "   " + HttpProxy.PROXY_PORT + "=<Proxy_Port>\n"
           + "   " + TRIPLEA_MEMORY_SET + "=true/false <did you set the xmx manually?>\n"
           + "   " + MAP_FOLDER + "=mapFolder"
           + "\n"
@@ -186,7 +172,7 @@ public class GameRunner {
     SwingUtilities.invokeLater(() -> LookAndFeel.setupLookAndFeel());
     showMainFrame();
     new Thread(() -> setupLogging(GameMode.SWING_CLIENT)).start();
-    setupProxies();
+    HttpProxy.setupProxies();
     new Thread(() -> checkForUpdates()).start();
     handleCommandLineArgs(args, getProperties(), GameMode.SWING_CLIENT);
   }
@@ -386,8 +372,8 @@ public class GameRunner {
     return new String[] {TRIPLEA_GAME_PROPERTY, TRIPLEA_SERVER_PROPERTY, TRIPLEA_CLIENT_PROPERTY, TRIPLEA_HOST_PROPERTY,
         TRIPLEA_PORT_PROPERTY, TRIPLEA_NAME_PROPERTY, TRIPLEA_SERVER_PASSWORD_PROPERTY, TRIPLEA_STARTED,
         LobbyServer.TRIPLEA_LOBBY_PORT_PROPERTY,
-        LOBBY_HOST, LOBBY_GAME_COMMENTS, LOBBY_GAME_HOSTED_BY, TRIPLEA_ENGINE_VERSION_BIN, PROXY_HOST, PROXY_PORT,
-        TRIPLEA_DO_NOT_CHECK_FOR_UPDATES, TRIPLEA_MEMORY_SET, MAP_FOLDER};
+        LOBBY_HOST, LOBBY_GAME_COMMENTS, LOBBY_GAME_HOSTED_BY, TRIPLEA_ENGINE_VERSION_BIN, HttpProxy.PROXY_HOST,
+        HttpProxy.PROXY_PORT, TRIPLEA_DO_NOT_CHECK_FOR_UPDATES, TRIPLEA_MEMORY_SET, MAP_FOLDER};
   }
 
   private static String getValue(final String arg) {
@@ -568,153 +554,6 @@ public class GameRunner {
     }
   }
 
-  private static void setupProxies() {
-    // System properties, not user pref
-    String proxyHostArgument = System.getProperty(PROXY_HOST);
-    String proxyPortArgument = System.getProperty(PROXY_PORT);
-    if (proxyHostArgument == null) {
-      // in case it was set by -D we also check this
-      proxyHostArgument = System.getProperty(HTTP_PROXYHOST);
-    }
-    if (proxyPortArgument == null) {
-      proxyPortArgument = System.getProperty(HTTP_PROXYPORT);
-    }
-    // arguments should override and set user preferences
-    String proxyHost = null;
-    if (proxyHostArgument != null && proxyHostArgument.trim().length() > 0) {
-      proxyHost = proxyHostArgument;
-    }
-    String proxyPort = null;
-    if (proxyPortArgument != null && proxyPortArgument.trim().length() > 0) {
-      try {
-        Integer.parseInt(proxyPortArgument);
-        proxyPort = proxyPortArgument;
-      } catch (final NumberFormatException e) {
-        ClientLogger.logQuietly(e);
-      }
-    }
-    if (proxyHost != null || proxyPort != null) {
-      setProxy(proxyHost, proxyPort, ProxyChoice.USE_USER_PREFERENCES);
-    }
-    final Preferences pref = Preferences.userNodeForPackage(GameRunner.class);
-    final ProxyChoice choice = ProxyChoice.valueOf(pref.get(PROXY_CHOICE, ProxyChoice.NONE.toString()));
-    if (choice == ProxyChoice.USE_SYSTEM_SETTINGS) {
-      setToUseSystemProxies();
-    } else if (choice == ProxyChoice.USE_USER_PREFERENCES) {
-      final String host = pref.get(PROXY_HOST, "");
-      final String port = pref.get(PROXY_PORT, "");
-      if (host.trim().length() > 0) {
-        System.setProperty(HTTP_PROXYHOST, host);
-      }
-      if (port.trim().length() > 0) {
-        System.setProperty(HTTP_PROXYPORT, port);
-      }
-    }
-  }
-
-  public static void setProxy(final String proxyHost, final String proxyPort, final ProxyChoice proxyChoice) {
-    final Preferences pref = Preferences.userNodeForPackage(GameRunner.class);
-    final ProxyChoice choice;
-    if (proxyChoice != null) {
-      choice = proxyChoice;
-      pref.put(PROXY_CHOICE, proxyChoice.toString());
-    } else {
-      choice = ProxyChoice.valueOf(pref.get(PROXY_CHOICE, ProxyChoice.NONE.toString()));
-    }
-    if (proxyHost != null && proxyHost.trim().length() > 0) {
-      // user pref, not system properties
-      pref.put(PROXY_HOST, proxyHost);
-      if (choice == ProxyChoice.USE_USER_PREFERENCES) {
-        System.setProperty(HTTP_PROXYHOST, proxyHost);
-      }
-    }
-    if (proxyPort != null && proxyPort.trim().length() > 0) {
-      try {
-        Integer.parseInt(proxyPort);
-        // user pref, not system properties
-        pref.put(PROXY_PORT, proxyPort);
-        if (choice == ProxyChoice.USE_USER_PREFERENCES) {
-          System.setProperty(HTTP_PROXYPORT, proxyPort);
-        }
-      } catch (final NumberFormatException e) {
-        ClientLogger.logQuietly(e);
-      }
-    }
-    if (choice == ProxyChoice.NONE) {
-      System.clearProperty(HTTP_PROXYHOST);
-      System.clearProperty(HTTP_PROXYPORT);
-    } else if (choice == ProxyChoice.USE_SYSTEM_SETTINGS) {
-      setToUseSystemProxies();
-    }
-    if (proxyHost != null || proxyPort != null || proxyChoice != null) {
-      try {
-        pref.flush();
-        pref.sync();
-      } catch (final BackingStoreException e) {
-        ClientLogger.logQuietly(e);
-      }
-    }
-  }
-
-  private static void setToUseSystemProxies() {
-    final String javaNetUseSystemProxies = "java.net.useSystemProxies";
-    System.setProperty(javaNetUseSystemProxies, "true");
-    List<Proxy> proxyList = null;
-    try {
-      final ProxySelector def = ProxySelector.getDefault();
-      if (def != null) {
-        proxyList = def.select(new URI("http://sourceforge.net/"));
-        ProxySelector.setDefault(null);
-        if (proxyList != null && !proxyList.isEmpty()) {
-          final Proxy proxy = proxyList.get(0);
-          final InetSocketAddress address = (InetSocketAddress) proxy.address();
-          if (address != null) {
-            final String host = address.getHostName();
-            final int port = address.getPort();
-            System.setProperty(HTTP_PROXYHOST, host);
-            System.setProperty(HTTP_PROXYPORT, Integer.toString(port));
-            System.setProperty(PROXY_HOST, host);
-            System.setProperty(PROXY_PORT, Integer.toString(port));
-          } else {
-            System.clearProperty(HTTP_PROXYHOST);
-            System.clearProperty(HTTP_PROXYPORT);
-            System.clearProperty(PROXY_HOST);
-            System.clearProperty(PROXY_PORT);
-          }
-        }
-      } else {
-        final String host = System.getProperty(PROXY_HOST);
-        final String port = System.getProperty(PROXY_PORT);
-        if (host == null) {
-          System.clearProperty(HTTP_PROXYHOST);
-        } else {
-          System.setProperty(HTTP_PROXYHOST, host);
-        }
-        if (port == null) {
-          System.clearProperty(HTTP_PROXYPORT);
-        } else {
-          try {
-            Integer.parseInt(port);
-            System.setProperty(HTTP_PROXYPORT, port);
-          } catch (final NumberFormatException nfe) {
-            // nothing
-          }
-        }
-      }
-    } catch (final Exception e) {
-      ClientLogger.logQuietly(e);
-    } finally {
-      System.setProperty(javaNetUseSystemProxies, "false");
-    }
-  }
-
-  public static void addProxy(final HostConfiguration config) {
-    final String host = System.getProperty(HTTP_PROXYHOST);
-    final String port = System.getProperty(HTTP_PROXYPORT, "-1");
-    if (host != null && host.trim().length() > 0) {
-      config.setProxy(host, Integer.valueOf(port));
-    }
-  }
 
   public static boolean getDelayedParsing() {
     final Preferences pref = Preferences.userNodeForPackage(GameRunner.class);
