+++ src/games/strategy/engine/random/PropertiesDiceRoller.java
@@ -19,10 +19,10 @@ import org.apache.commons.httpclient.methods.PostMethod;
 import games.strategy.debug.ClientLogger;
 import games.strategy.engine.ClientContext;
 import games.strategy.engine.ClientFileSystemHelper;
-import games.strategy.engine.framework.GameRunner;
 import games.strategy.engine.framework.startup.ui.editors.DiceServerEditor;
 import games.strategy.engine.framework.startup.ui.editors.EditorPanel;
 import games.strategy.engine.framework.startup.ui.editors.IBean;
+import games.strategy.engine.framework.system.HttpProxy;
 
 /**
  * A pbem dice roller that reads its configuration from a properties file
@@ -131,8 +131,7 @@ public class PropertiesDiceRoller implements IRemoteDiceServer {
       }
       final HostConfiguration config = client.getHostConfiguration();
       config.setHost(host, port);
-      // add the proxy
-      GameRunner.addProxy(config);
+      HttpProxy.addProxy(config);
       client.executeMethod(post);
       final String result = post.getResponseBodyAsString();
       return result;
